Move trends report controller/route to mixins.
